Sun, Oct 1, 2023, 6:25 PM -04:00

+++ chapter1/population/population.c
@@ -4,7 +4,7 @@
 
 int main(void)
 {
-    float Start
+    int Start
     int Start = get_int("Starting Population: ");
     do
     {
